# Schems/swagger.yaml
openapi: 3.0.4
info:
  title: OpenVPN Gate Monitor API
  version: v1
paths:
...
  /api/quota-plans/get-all:
    post:
      tags:
        - QuotaPlan
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/GetQuotaPlansRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/GetQuotaPlansRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/GetQuotaPlansRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/GetQuotaPlansRequest'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/QuotaPlansResponseApiResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/QuotaPlansResponseApiResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/QuotaPlansResponseApiResponse'
  '/api/quota-plans/get/{id}':
    get:
      tags:
        - QuotaPlan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/QuotaPlanResponseApiResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/QuotaPlanResponseApiResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/QuotaPlanResponseApiResponse'
  /api/quota-plans/create:
    post:
      tags:
        - QuotaPlan
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/CreateOrUpdateQuotaPlanRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrUpdateQuotaPlanRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/CreateOrUpdateQuotaPlanRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/CreateOrUpdateQuotaPlanRequest'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/QuotaPlanResponseApiResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/QuotaPlanResponseApiResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/QuotaPlanResponseApiResponse'
  /api/quota-plans/update:
    put:
      tags:
        - QuotaPlan
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/CreateOrUpdateQuotaPlanRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrUpdateQuotaPlanRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/CreateOrUpdateQuotaPlanRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/CreateOrUpdateQuotaPlanRequest'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/StringApiResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/StringApiResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/StringApiResponse'
  '/api/quota-plans/delete/{id}':
    delete:
      tags:
        - QuotaPlan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/StringApiResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/StringApiResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/StringApiResponse'
...
  /api/tgbot-incoming-message-logs/add:
    post:
      tags:
        - TelegramBotIncomingMessageLog
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/AddMessageRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/AddMessageRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/AddMessageRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/AddMessageRequest'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/AddMessageResponseApiResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/AddMessageResponseApiResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/AddMessageResponseApiResponse'
...
  /api/tgbot-incoming-message-logs/get-by-id:
    get:
      tags:
        - TelegramBotIncomingMessageLog
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/GetByIdMessageRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/GetByIdMessageRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/GetByIdMessageRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/GetByIdMessageRequest'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/GetByIdMessageResponseApiResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/GetByIdMessageResponseApiResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/GetByIdMessageResponseApiResponse'
...
components:
  schemas:
...
    AddMessageRequest:
      type: object
      properties:
        message:
          $ref: '#/components/schemas/MessageDto'
      additionalProperties: false
    AddMessageResponse:
      type: object
      properties:
        message:
          $ref: '#/components/schemas/MessageDto'
      additionalProperties: false
    AddMessageResponseApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
          nullable: true
        data:
          $ref: '#/components/schemas/AddMessageResponse'
      additionalProperties: false
...
    CreateOrUpdateQuotaPlanRequest:
      required:
        - name
        - overlimitAction
      type: object
      properties:
        id:
          maximum: 2147483647
          minimum: 0
          type: integer
          format: int32
        name:
          maxLength: 64
          minLength: 1
          type: string
        description:
          maxLength: 256
          type: string
          nullable: true
        dailyQuotaBytes:
          maximum: 9223372036854776000
          minimum: 0
          type: integer
          format: int64
          nullable: true
        monthlyQuotaBytes:
          maximum: 9223372036854776000
          minimum: 0
          type: integer
          format: int64
          nullable: true
        upKbps:
          maximum: 2147483647
          minimum: 0
          type: integer
          format: int32
          nullable: true
        downKbps:
          maximum: 2147483647
          minimum: 0
          type: integer
          format: int32
          nullable: true
        overlimitAction:
          $ref: '#/components/schemas/QuotaOverlimitAction'
        throttleUpKbps:
          maximum: 2147483647
          minimum: 0
          type: integer
          format: int32
          nullable: true
        throttleDownKbps:
          maximum: 2147483647
          minimum: 0
          type: integer
          format: int32
          nullable: true
        isActive:
          type: boolean
        isDefault:
          type: boolean
      additionalProperties: false
...
    GetByIdMessageRequest:
      required:
        - id
      type: object
      properties:
        id:
          maximum: 2147483647
          minimum: 1
          type: integer
          format: int32
      additionalProperties: false
    GetByIdMessageResponse:
      type: object
      properties:
        messages:
          $ref: '#/components/schemas/MessageDto'
      additionalProperties: false
    GetByIdMessageResponseApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
          nullable: true
        data:
          $ref: '#/components/schemas/GetByIdMessageResponse'
      additionalProperties: false
...
    GetQuotaPlansRequest:
      type: object
      properties:
        includeInactive:
          type: boolean
      additionalProperties: false
...
    MessageDto:
      required:
        - telegramId
      type: object
      properties:
        id:
          type: integer
          format: int32
        telegramId:
          type: integer
          format: int64
        username:
          type: string
          nullable: true
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        messageText:
          type: string
          nullable: true
        fileType:
          type: string
          nullable: true
        fileId:
          type: string
          nullable: true
        fileName:
          type: string
          nullable: true
        fileSize:
          type: integer
          format: int64
          nullable: true
        filePath:
          type: string
          nullable: true
        receivedAt:
          type: string
          format: date-time
        createDate:
          type: string
          format: date-time
        lastUpdate:
          type: string
          format: date-time
      additionalProperties: false
...
    QuotaOverlimitAction:
      enum:
        - 0
        - 1
        - 2
        - 3
      type: integer
      format: int32
    QuotaPlanDto:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        dailyQuotaBytes:
          type: integer
          format: int64
          nullable: true
        monthlyQuotaBytes:
          type: integer
          format: int64
          nullable: true
        upKbps:
          type: integer
          format: int32
          nullable: true
        downKbps:
          type: integer
          format: int32
          nullable: true
        overlimitAction:
          $ref: '#/components/schemas/QuotaOverlimitAction'
        throttleUpKbps:
          type: integer
          format: int32
          nullable: true
        throttleDownKbps:
          type: integer
          format: int32
          nullable: true
        isActive:
          type: boolean
        isDefault:
          type: boolean
      additionalProperties: false
    QuotaPlanResponse:
      type: object
      properties:
        quotaPlan:
          $ref: '#/components/schemas/QuotaPlanDto'
      additionalProperties: false
    QuotaPlanResponseApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
          nullable: true
        data:
          $ref: '#/components/schemas/QuotaPlanResponse'
      additionalProperties: false
    QuotaPlansResponse:
      type: object
      properties:
        quotaPlans:
          type: array
          items:
            $ref: '#/components/schemas/QuotaPlanDto'
          nullable: true
      additionalProperties: false
    QuotaPlansResponseApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
          nullable: true
        data:
          $ref: '#/components/schemas/QuotaPlansResponse'
      additionalProperties: false
...
    StringApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
          nullable: true
        data:
          type: string
          nullable: true
      additionalProperties: false
...
  securitySchemes:
    Bearer:
      type: http
      description: 'Enter the token in the format ''Bearer {your_token}'''
      scheme: Bearer
      bearerFormat: JWT
security:
  - Bearer: [ ]